refactor CreateNewStrategyModal
